+++ intelmq/lib/upgrades.py
@@ -18,6 +18,7 @@ __all__ = ['v100_dev7_modify_syntax',
            'v200_defaults_ssl_ca_certificate',
            'v111_defaults_process_manager',
            'v202_fixes',
+           'v210_deprecations',
            ]
 
 
@@ -239,6 +240,22 @@ def v202_fixes(defaults, runtime, dry_run):
     return changed, defaults, runtime
 
 
+def v210_deprecations(defaults, runtime, dry_run):
+    """
+    Migrating RT collector's `unzip_attachment` to `extract_files`.
+    """
+    changed = None
+    for bot_id, bot in runtime.items():
+        if bot["module"] == "intelmq.bots.collectors.rt.collector_rt":
+            if "unzip_attachment" not in bot["parameters"]:
+                continue
+            if "extract_files" not in bot["parameters"]:
+                bot["parameters"]["extract_files"] = bot["parameters"]["unzip_attachment"]
+            del bot["parameters"]["unzip_attachment"]
+            changed = True
+    return changed, defaults, runtime
+
+
 UPGRADES = OrderedDict([
     ((1, 0, 0, 'dev7'), (v100_dev7_modify_syntax, )),
     ((1, 1, 0), (v110_shadowserver_feednames, v110_deprecations)),
@@ -248,4 +265,5 @@ UPGRADES = OrderedDict([
                  v200_defaults_ssl_ca_certificate)),
     ((2, 0, 1), ()),
     ((2, 0, 2), (v202_fixes, )),
+    ((2, 1, 0), (v210_deprecations, )),
 ])